+++ app/src/main/java/com/lemon/mcdevmanager/ui/fragPage/OverviewFragPage.kt
@@ -1,0 +1,29 @@
+package com.lemon.mcdevmanager.ui.fragPage
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import com.lemon.mcdevmanager.ui.theme.AppTheme
+import com.lemon.mcdevmanager.ui.theme.MCDevManagerTheme
+
+@Composable
+fun OverviewFragPage() {
+
+}
+
+@Composable
+@Preview(showBackground = true, showSystemUi = true)
+fun OverviewFragPagePreview() {
+    MCDevManagerTheme {
+        Box(
+            modifier = Modifier
+                .fillMaxSize()
+                .background(AppTheme.colors.background)
+        ) {
+            OverviewFragPage()
+        }
+    }
+}
