+++ app/src/main/java/com/lemon/mcdevmanager/ui/theme/Color.kt
@@ -8,7 +8,7 @@ val Pink200 = Color(0xFFEF9A9A)
 
 val Purple40 = Color(0xFF6650a4)
 val PurpleGrey40 = Color(0xFF625b71)
-val Pink40 = Color(0xFF7D5260)
+val Pink40 = Color(0xFFE57373)
 
 val TextDay = Color(0xFF121212)
 val TextNight = Color(0xFFCCCCCC)
