+++ app/src/main/java/com/lemon/mcdevmanager/viewModel/SplashViewModel.kt
@@ -6,6 +6,7 @@ import com.lemon.mcdevmanager.data.common.LOGIN_PAGE
 import com.lemon.mcdevmanager.data.common.MAIN_PAGE
 import com.lemon.mcdevmanager.data.database.database.GlobalDataBase
 import com.lemon.mcdevmanager.data.global.AppContext
+import com.orhanobut.logger.Logger
 import com.zj.mvi.core.SharedFlowEvents
 import com.zj.mvi.core.setEvent
 import kotlinx.coroutines.Dispatchers
